refactor: general refactor to adhere the tech challenge structure
